# Speed up copy in MapVector & ArrayVector