Summary:
Today, the implementations of copy in MapVector and ArrayVector operate sequentially resizing the child
Vectors one map/array at a time and copying data one map/array at a time.

These leads to unnecessary overhead and can be optimized especially in the case where the maps/arrays
being copied point to sequential elements in the child Vectors.

This diff makes 3 changes to these functions:
* We do a pass over all the maps/arrays to be copied to determine the total amount we need to grow the
child vectors by and does a single resize.
* We collect the ranges of elements in the child vectors to copy and merge adjacent ranges that appear
consecutively.
* We special case copying a single map/array (these changes added significant overhead >10% in the
case of copying a single map/array, and special casing this makes the change neutral in this case).

On a side note I also considered making a map of ranges keyed on the end offset, but this would not work
in the rare case that the maps/arrays pointed to two or more overlapping ranges that ended at the same \
offset, so I relied on just checking adjacent ranges (this seems like the more likely case anyway).

I added a benchmark to verify the changes yielded improvements.

Original results:
```
============================================================================
velox/vector/benchmarks/CopyBenchmark.cpp     relative  time/iter   iters/s
============================================================================
copyArray                                                  19.75ns    50.64M
copyArrayWithNulls                                         22.74ns    43.98M
copyArrayDictionaryEncoded                                 44.73ns    22.36M
copyArrayOneAtATime                                        60.52ns    16.52M
copyArrayTwoAtATime                                        40.43ns    24.73M
copyMap                                                    36.37ns    27.49M
copyMapWithNulls                                           38.84ns    25.75M
copyMapDictionaryEncoded                                   88.00ns    11.36M
copyMapOneAtATime                                          76.89ns    13.00M
copyMapTwoAtATime                                          57.24ns    17.47M
```

After these changes:
```
============================================================================
velox/vector/benchmarks/CopyBenchmark.cpp     relative  time/iter   iters/s
============================================================================
copyArray                                                   5.57ns   179.60M
copyArrayWithNulls                                          8.54ns   117.14M
copyArrayDictionaryEncoded                                 28.09ns    35.60M
copyArrayOneAtATime                                        57.87ns    17.28M
copyArrayTwoAtATime                                        39.23ns    25.49M
copyMap                                                     8.19ns   122.04M
copyMapWithNulls                                           10.40ns    96.15M
copyMapDictionaryEncoded                                   53.80ns    18.59M
copyMapOneAtATime                                          73.76ns    13.56M
copyMapTwoAtATime                                          46.20ns    21.65M
```

For copying maps/arrays laid out consecutively, there's a ~4x improvement.  Even for copying
maps/arrays laid out non-consecutively (dictionary encoded) there's a ~40% improvement.

For copying a single map/array or two arrays, there's no change.

For copying two maps we see a 20% improvement.

Differential Revision: D38248521